+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-05-11
+
+### Added
+
+- New AI Weighed
+- Function to run AI without rendering
+- getSize method for board
+- function to compare to arrays
+
+### Changed
+
+- Reproducible shuffling
+
 ## [1.0.0] - 2023-05-04
 
 ### Changed